requirements.txt
Flask==3.0.3

# --- Database ---
SQLAlchemy
PyMySQL
python-dotenv

# --- Optional for API (if using FastAPI) ---
FastAPI
uvicorn